# Conflicts:
#	.idea/misc.xml
#	app/src/main/AndroidManifest.xml
#	app/src/main/res/layout/activity_add_event.xml
#	app/src/main/res/layout/fragment_calendar.xml

+++ app/src/main/java/com/example/ipcalink/calendar/RecyclerViewAdapter.kt
@@ -3,12 +3,14 @@ package com.example.ipcalink.calendar
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.content.Intent
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContentProviderCompat.requireContext
+import androidx.core.content.ContextCompat.startActivity
 import androidx.recyclerview.widget.RecyclerView
 import com.example.ipcalink.R
 import com.example.ipcalink.calendar.CalendarHelper.DateFormater
@@ -24,11 +26,12 @@ import com.kizitonwose.calendarview.CalendarView
 import kotlinx.coroutines.NonDisposableHandle.parent
 import java.text.SimpleDateFormat
 import java.time.LocalDate
+import java.time.temporal.ChronoUnit
 import java.util.*
 import java.util.regex.Pattern
 import kotlin.collections.ArrayList
 
-class RecyclerViewAdapter internal constructor(rl: MutableList<Events>, map : MutableMap<LocalDate, List<Events>>, b : FragmentCalendarBinding) : RecyclerView.Adapter<RecyclerViewAdapter.RecyclerViewHolder>() {
+class RecyclerViewAdapter internal constructor(rl: MutableList<Events>, map : MutableMap<LocalDate, List<Events>>, b : FragmentCalendarBinding, c : Context) : RecyclerView.Adapter<RecyclerViewAdapter.RecyclerViewHolder>() {
 
     private val userUID = Firebase.auth.uid
     private val dbFirebase = Firebase.firestore
@@ -37,6 +40,7 @@ class RecyclerViewAdapter internal constructor(rl: MutableList<Events>, map : Mu
     private val binding = b
 
     private val recyclerMap = map
+    private val context = c
 
     private var recyclerList = rl
 
@@ -72,6 +76,16 @@ class RecyclerViewAdapter internal constructor(rl: MutableList<Events>, map : Mu
                 .show()
         }*/
 
+        holder.itemView.setOnClickListener {
+
+            val intent = Intent(context, EditEventActivity::class.java)
+            //intent.putExtra("dayOfWeek", dayOfWeek.toString())
+            intent.putExtra("eventId", recyclerList[position].id)
+
+
+            startActivity(context, intent, null)
+        }
+
 
         holder.itemView.apply {
 
@@ -92,7 +106,7 @@ class RecyclerViewAdapter internal constructor(rl: MutableList<Events>, map : Mu
 
             holder.title?.text = currentItem.title
             holder.duration?.text = "$startTime-$endTime"
-            holder.chatName?.text = calendarSharedPreferences(holder.itemView.context).currentChatId
+            holder.chatName?.text = calendarSharedPreferences(holder.itemView.context).currentChatName
         }
     }
 
@@ -112,20 +126,60 @@ class RecyclerViewAdapter internal constructor(rl: MutableList<Events>, map : Mu
 
     private fun deleteEvent(event: Events, currentChatId : String?, currentChatName : String?) {
 
+        binding.calendar.notifyCalendarChanged()
 
         val startDate = getDate(event.startDate!!.seconds * 1000, "yyyy-MM-dd'T'HH:mm:ss.SSS")
         val endDate = getDate(event.endDate!!.seconds * 1000, "yyyy-MM-dd'T'HH:mm:ss.SSS")
 
 
         val localStartDate = LocalDate.parse(DateFormater(startDate))
-        binding.calendar.notifyDateChanged(localStartDate)
+        val localendDate = LocalDate.parse(DateFormater(endDate))
 
+        val dayDiff = ChronoUnit.DAYS.between(localStartDate, localendDate)
 
+        var date = localStartDate
 
-        recyclerMap[localStartDate] = recyclerMap[localStartDate].orEmpty().minus(event)
-        updateAdapterForDate(localStartDate)
+        var i = 0
 
-        deleteEvents(event, currentChatId, currentChatName)
+        while (i <= dayDiff) {
+            binding.calendar.notifyDateChanged(date)
+
+            date?.let {
+                recyclerMap[date] = recyclerMap[date].orEmpty().minus(Events(event.id, event.title, event.description, event.sendDate, event.senderId, event.startDate, event.endDate))
+                updateAdapterForDate(it)
+            }
+
+            date = date.plusDays(1)
+
+            i++
+        }
+
+        AlertDialog.Builder(context)
+            .setMessage("Delete this event?")
+            .setPositiveButton("Confirm") { _, _ ->
+                deleteEvents(event, currentChatId, currentChatName)
+            }
+            .setNegativeButton("Cancel") { _, _ ->
+                recyclerMap.clear()
+
+                date = localStartDate
+                i = 0
+
+                while (i <= dayDiff) {
+                    binding.calendar.notifyDateChanged(date)
+
+                    date?.let {
+                        recyclerMap[date] = recyclerMap[date].orEmpty().plus(Events(event.id, event.title, event.description, event.sendDate, event.senderId, event.startDate, event.endDate))
+                        updateAdapterForDate(it)
+                    }
+
+                    date = date.plusDays(1)
+
+                    i++
+                }
+            }.show()
+
+        //deleteEvents(event, currentChatId, currentChatName)
     }
 
     private fun deleteEvents(event : Events, currentChatId : String?, currentChatName : String?) {
@@ -159,7 +213,7 @@ class RecyclerViewAdapter internal constructor(rl: MutableList<Events>, map : Mu
         recyclerList.clear()
 
 
-        recyclerList.addAll(recyclerMap[date].orEmpty())
+        recyclerList.removeAll(recyclerMap[date].orEmpty())
 
         //val month = monthTitleFormatter.format(date.month)
         //binding.textViewMonth.text = month
